# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jack — AI Trade Intelligence Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /graph/candidates:
    get:
      operationId: listKnowledgeCandidates
      tags: [graph]
      summary: List mentor-concept candidates queued for review (read-only)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, accepted, rejected, merged, archived, restored]
            default: pending
          description: >-
            pending is publicly readable; every other status (including
            archived — mentor-withdrawn concepts held out of the live graph,
            and restored — archived concepts a reviewer re-minted into the
            graph) requires an admin session.
      responses:
        "200":
          description: Queued knowledge candidates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeCandidateList"
  /graph/mentor-contributions:
    get:
      operationId: getMentorContributions
      tags: [graph]
      summary: Per-mentor contribution counts for reviewer trust calibration (admin only)
      description: >-
        Read-only aggregation over mentor→concept provenance edges and the
        knowledge-candidate history. Reviewers use it to gauge a mentor's
        overall track record when weighing a borderline candidate. Admin-gated
        like the rest of the resolved-candidate surface.
      responses:
        "200":
          description: Per-mentor contribution counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorContributionList"
        "401":
          description: Admin session required
  /graph/candidates/{id}/resolve:
    post:
      operationId: resolveKnowledgeCandidate
      tags: [graph]
      summary: Resolve a pending knowledge candidate — Accept / Merge / Reject (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateResolutionInput"
      responses:
        "200":
          description: The resolved candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeCandidate"
        "400":
          description: Invalid resolution (missing reason/target, or bad target node)
        "404":
          description: Candidate not found
        "409":
          description: >-
            Conflict — either the candidate was already resolved with a
            different outcome (code=already_resolved), or the requested target
            concept no longer exists in the live graph and nothing confidently
            replaces it (code=target_gone; the candidate STAYS pending and
            bestMatches carries fresh near matches to choose from).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateResolutionConflict"
  /graph/nodes/{id}/verification:
    patch:
      operationId: setNodeVerification
      tags: [graph]
      summary: Set the human verification status of a distilled knowledge node (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationUpdate"
      responses:
        "200":
          description: Updated knowledge node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeNode"
  /graph/nodes/{id}/restore-evidence:
    post:
      operationId: restoreWithdrawnEvidence
      tags: [graph]
      summary: >-
        Clear a reviewed withdrawn-evidence entry from a concept's provenance
        (admin only). Idempotent — a missing entry is a no-op success.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreEvidenceInput"
      responses:
        "200":
          description: Updated knowledge node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeNode"
        "400":
          description: Invalid node id or body
        "404":
          description: No distilled knowledge node with that id
...
components:
  schemas:
...
    VerificationUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [verified, rejected, unverified]
    RestoreEvidenceInput:
      type: object
      required: [videoId]
      properties:
        videoId:
          type: string
          description: >-
            The source video whose withdrawn-evidence entry a reviewer has
            reviewed and wants cleared from this concept's provenance history.
    KnowledgeNode:
      type: object
      required:
        [id, kind, label, description, confidence, verificationStatus, meta, createdAt, updatedAt]
      properties:
        id:
          type: string
        kind:
          type: string
          enum:
            [
              core,
              topic,
              competency,
              video,
              mentor,
              concept,
              tool,
              equipment,
              material,
              procedure,
              hazard,
              slang,
              certification,
              standard,
              regional_term,
            ]
        label:
          type: string
        trade:
          type: ["string", "null"]
        refId:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        confidence:
          type: ["number", "null"]
        verificationStatus:
          type: string
        meta:
          type: object
          additionalProperties: true
        createdAt:
          type: string
        updatedAt:
          type: string
...
    KnowledgeCandidateMatch:
      type: object
      required: [nodeId, label, similarity]
      properties:
        nodeId:
          type: string
        label:
          type: string
        similarity:
          type: number
        validity:
          type: string
          enum: [live, redirected, gone]
          description: >-
            Read-time annotation of whether this recorded match still exists in
            the live graph: live — usable as-is; redirected — absorbed into
            another concept (see currentNodeId/currentLabel); gone — no longer
            exists and left no redirect trail.
        currentNodeId:
          type: ["string", "null"]
          description: The node this match currently resolves to (itself when live, the survivor when redirected).
        currentLabel:
          type: ["string", "null"]
    KnowledgeCandidate:
      type: object
      required: [id, status, title, category, bestMatches]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, merged, archived, restored]
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        trade:
          type: ["string", "null"]
        confidence:
          type: ["number", "null"]
        competencyCode:
          type: ["string", "null"]
        mentorProfileId:
          type: ["string", "null"]
        mentorName:
          type: ["string", "null"]
        answerId:
          type: ["string", "null"]
        sessionId:
          type: ["string", "null"]
        bestMatches:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeCandidateMatch"
        createdAt:
          type: ["string", "null"]
        resolvedTargetId:
          type: ["string", "null"]
          description: The canonical concept node reinforced by an accept/merge resolution.
        resolutionReason:
          type: ["string", "null"]
          description: The reviewer's reason, recorded on reject.
        resolvedAt:
          type: ["string", "null"]
        requestedTargetId:
          type: ["string", "null"]
          description: The target the reviewer originally asked for (accept/merge only).
        redirectReason:
          type: ["string", "null"]
          description: >-
            Why the recorded resolvedTargetId differs from requestedTargetId —
            set when the requested concept was merged away or re-matched by
            content at resolution time; null when the target was used as-is.
    CandidateResolutionInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [accept, merge, reject, restore, rearchive, reopen]
          description: >-
            accept — reinforce the top best-match concept; merge — reinforce the
            reviewer-chosen targetNodeId; reject — discard with a required reason;
            restore — re-mint an archived (mentor-withdrawn) concept as
            attribution-free unverified knowledge; rearchive — undo a restore,
            demoting the curated concept back to an archived candidate (removing
            the sourceless node from the live graph, or dropping only the
            reviewer's curated vouch if a video/mentor re-taught it meanwhile);
            reopen — undo a resolved decision, returning the candidate to the
            pending queue for a fresh review (rejected, accepted, or merged
            candidates). Reject wrote no graph edge so its reopen is side-effect-
            free; accept/merge first withdraw this candidate's contribution from
            the mentor→concept edge (decrementing its weight, deleting the edge
            when its last answer is removed) and re-evaluate the concept node.
            A scrubbed (withdrawn-mentor) candidate can never be reopened.
        targetNodeId:
          type: string
          description: The existing concept to merge into (required for merge).
        reason:
          type: string
          description: Why the candidate was rejected (required for reject).
    CandidateResolutionConflict:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [already_resolved, target_gone]
        bestMatches:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeCandidateMatch"
          description: Fresh near matches from the live graph (target_gone only).
    KnowledgeCandidateList:
      type: object
      required: [candidates, total]
      properties:
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeCandidate"
        total:
          type: integer
    MentorContribution:
      type: object
      required:
        [mentorProfileId, conceptsCreated, conceptsReinforced, accepted, rejected, pending]
      properties:
        mentorProfileId:
          type: string
        conceptsCreated:
          type: integer
          description: >-
            Live concepts this mentor is the sole provenance source for
            (no other video or mentor corroborates them).
        conceptsReinforced:
          type: integer
          description: >-
            Live concepts this mentor co-sources alongside other videos or
            mentors (strengthening a shared canonical concept).
        accepted:
          type: integer
          description: Review candidates from this mentor accepted or merged into the graph.
        rejected:
          type: integer
          description: Review candidates from this mentor rejected.
        pending:
          type: integer
          description: Review candidates from this mentor still awaiting a decision.
    MentorContributionList:
      type: object
      required: [contributions, total]
      properties:
        contributions:
          type: array
          items:
            $ref: "#/components/schemas/MentorContribution"
        total:
          type: integer
...